+++ src/Indice.Features.Cases.AspNetCore/Workflows/Activities/GetCaseDetailsActivity.cs
@@ -37,6 +37,9 @@ namespace Indice.Features.Cases.Workflows.Activities
             // Run as systemic user, since this is a system activity for creating conditions at workflow
             var systemUser = Cases.Extensions.PrincipalExtensions.SystemUser();
             var @case = await _adminCaseService.GetCaseById(systemUser, CaseId.Value, IncludeAttachmentsData);
+            
+            // Convert CaseData to JObject so the workflow activities can use data without parsing.
+            @case.Data = Newtonsoft.Json.Linq.JObject.Parse(@case.DataAs<string>());
             Output = @case;
             context.LogOutputProperty(this, nameof(Output), Output);
             return Done(Output);